chat: consolidate item source files into single module
Merge aiCustomizationListItem, aiCustomizationItemSourceUtils,
aiCustomizationItemNormalizer, and providerCustomizationItemSource
into a single aiCustomizationItemSource.ts (~414 lines).

These four files form a tight linear dependency chain with one
external consumer (the list widget). Consolidating matches the
codebase convention (cf. testing explorerProjections/index.ts)
and reduces the aiCustomization directory from 22 to 19 files.

The promptsServiceCustomizationItemProvider remains separate as
a distinct adapter that bridges the core promptsService into the
provider-shaped pipeline for harnesses without an external provider.
